Move cumulative into separate module
